+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+## Version NEXT
+
+- Add *non-final* `FY2025` module.  Previous year values are used
+  for some thresholds and rates that are yet to be declared.
+
+
 ## Version 2024.1
 
 - The minimum supported version of GHC is now 9.2, due to a